refactor: stuff file stuff in file stuff file
There are some things that are a bit in-the-middley, but in general
there's a class of code about safely saving stuff to and loading stuff
from files (and validating them on the way) and a class of code about
calling cryptography to build, manipulate, and use certs. Split that up
so that only the latter is in cryptography_utils, and the former is in a
new file file_utils. Good de-bloating of cryptography_utils and more
importantly its tests
